project/requirements.txt
pyTelegramBotAPI
SQLAlchemy
python-dotenv